[worker] mcp split 2/6: extract shared helpers and response/warnings infrastructure
